Backend/app/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1

SQLAlchemy==2.0.36
psycopg[binary]==3.2.3

python-jose==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
python-multipart==0.0.12

pydantic==2.9.2
pydantic-settings==2.6.1

reportlab==4.2.5
email-validator==2.2.0